backend/requirements.txt
cobble==0.1.4
mammoth==1.9.0
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
pytesseract==0.3.13
